requirements.txt
asgiref==3.3.4
boto3==1.17.76
botocore==1.20.76
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-addthis==3.0.1
django-crispy-forms==1.11.2
django-jazzmin==2.4.7
django-storages==1.11.1
django-tinymce==3.3.0
fontawesome-free==5.15.3
idna==2.10
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
